docs: improve demo card suspense fallback style
